Increase MaldiPlateTable font size

+++ s2p-gui/src/main/java/es/uvigo/ei/sing/s2p/gui/UISettings.java
@@ -4,6 +4,6 @@ import java.awt.Color;
 
 public class UISettings {
 	public static final Color 	BG_COLOR 			= Color.WHITE;
-	public static final int 	FONT_SIZE 			= 14;
+	public static final float 	FONT_SIZE 			= 14;
 	public static final int 	FONT_SIZE_HEADER 	= 14;
 }
